Fix race condition in reading EngineEvents from tailed logs
We we're seeing the odd test fail because preview returned that len(summaryEvents) was 0. This was because everything about this code is async and didn't do any synchronisation so it was pure luck that the file was read, the message was sent to the channel, and then the channel goroutine read the message and appened it to summaryEvents in time for the len check to pass.

Now we close off the file watcher and wait for the channels and goroutines to flush and signal done before checking the length of summaryEvents.

+++ sdk/go/auto/stack.go
@@ -264,17 +264,19 @@ func (s *Stack) Preview(ctx context.Context, opts ...optpreview.Option) (Preview
 
 	var summaryEvents []apitype.SummaryEvent
 	eventChannel := make(chan events.EngineEvent)
-	go func(ch chan events.EngineEvent, events *[]apitype.SummaryEvent) {
+	eventsDone := make(chan bool)
+	go func() {
 		for {
 			event, ok := <-eventChannel
 			if !ok {
+				close(eventsDone)
 				return
 			}
 			if event.SummaryEvent != nil {
 				summaryEvents = append(summaryEvents, *event.SummaryEvent)
 			}
 		}
-	}(eventChannel, &summaryEvents)
+	}()
 
 	eventChannels := []chan<- events.EngineEvent{eventChannel}
 	eventChannels = append(eventChannels, preOpts.EventStreams...)
@@ -283,13 +285,17 @@ func (s *Stack) Preview(ctx context.Context, opts ...optpreview.Option) (Preview
 	if err != nil {
 		return res, errors.Wrap(err, "failed to tail logs")
 	}
-	defer cleanup(t, eventChannels)
+	defer t.Close()
 	args = append(args, "--event-log", t.Filename)
 
 	stdout, stderr, code, err := s.runPulumiCmdSync(ctx, preOpts.ProgressStreams /* additionalOutput */, args...)
 	if err != nil {
 		return res, newAutoError(errors.Wrap(err, "failed to run preview"), stdout, stderr, code)
 	}
+
+	// Close the file watcher wait for all events to send
+	t.Close()
+	<-eventsDone
 
 	if len(summaryEvents) == 0 {
 		return res, newAutoError(errors.New("failed to get preview summary"), stdout, stderr, code)
@@ -367,7 +373,7 @@ func (s *Stack) Up(ctx context.Context, opts ...optup.Option) (UpResult, error) 
 		if err != nil {
 			return res, errors.Wrap(err, "failed to tail logs")
 		}
-		defer cleanup(t, eventChannels)
+		defer t.Close()
 		args = append(args, "--event-log", t.Filename)
 	}
 
@@ -444,7 +450,7 @@ func (s *Stack) Refresh(ctx context.Context, opts ...optrefresh.Option) (Refresh
 		if err != nil {
 			return res, errors.Wrap(err, "failed to tail logs")
 		}
-		defer cleanup(t, eventChannels)
+		defer t.Close()
 		args = append(args, "--event-log", t.Filename)
 	}
 
@@ -515,7 +521,7 @@ func (s *Stack) Destroy(ctx context.Context, opts ...optdestroy.Option) (Destroy
 		if err != nil {
 			return res, errors.Wrap(err, "failed to tail logs")
 		}
-		defer cleanup(t, eventChannels)
+		defer t.Close()
 		args = append(args, "--event-log", t.Filename)
 	}
 
@@ -983,7 +989,56 @@ func (s *languageRuntimeServer) InstallDependencies(
 	return nil
 }
 
-func tailLogs(command string, receivers []chan<- events.EngineEvent) (*tail.Tail, error) {
+type fileWatcher struct {
+	Filename  string
+	tail      *tail.Tail
+	receivers []chan<- events.EngineEvent
+	done      chan bool
+}
+
+func watchFile(path string, receivers []chan<- events.EngineEvent) (*fileWatcher, error) {
+	t, err := tail.TailFile(path, tail.Config{
+		Follow: true,
+		Logger: tail.DiscardingLogger,
+	})
+	if err != nil {
+		return nil, err
+	}
+	done := make(chan bool)
+	go func(tailedLog *tail.Tail) {
+		for line := range tailedLog.Lines {
+			if line.Err != nil {
+				for _, r := range receivers {
+					r <- events.EngineEvent{Error: line.Err}
+				}
+				continue
+			}
+			var e apitype.EngineEvent
+			err = json.Unmarshal([]byte(line.Text), &e)
+			if err != nil {
+				for _, r := range receivers {
+					r <- events.EngineEvent{Error: err}
+				}
+				continue
+			}
+			for _, r := range receivers {
+				r <- events.EngineEvent{EngineEvent: e}
+			}
+		}
+		for _, r := range receivers {
+			close(r)
+		}
+		close(done)
+	}(t)
+	return &fileWatcher{
+		Filename:  t.Filename,
+		tail:      t,
+		receivers: receivers,
+		done:      done,
+	}, nil
+}
+
+func tailLogs(command string, receivers []chan<- events.EngineEvent) (*fileWatcher, error) {
 	logDir, err := ioutil.TempDir("", fmt.Sprintf("automation-logs-%s-", command))
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to create logdir")
@@ -998,11 +1053,19 @@ func tailLogs(command string, receivers []chan<- events.EngineEvent) (*tail.Tail
 	return t, nil
 }
 
-func cleanup(t *tail.Tail, channels []chan<- events.EngineEvent) {
-	logDir := filepath.Dir(t.Filename)
-	t.Cleanup()
-	os.RemoveAll(logDir)
-	for _, ch := range channels {
-		close(ch)
+func (fw *fileWatcher) Close() {
+	if fw.tail == nil {
+		return
 	}
+
+	// Tell the watcher to end on next EoF
+	fw.tail.StopAtEOF()
+	// Wait for our done event
+	<-fw.done
+	// Clean up
+	logDir := filepath.Dir(fw.tail.Filename)
+	fw.tail.Cleanup()
+	os.RemoveAll(logDir)
+	// set to nil so we can safely close again in defer
+	fw.tail = nil
 }
